Factor out abstract mojo on deployment level
